initial transfer plus linting
