--- paloalto/.gitignore
@@ -1,2 +1,0 @@
-ROADMAP.md
-*.sh
